add sample for eventhub batch
