+++ eventhubs/azure-spring-cloud-stream-binder-eventhubs/eventhubs-binder/src/main/java/com/azure/spring/sample/eventhubs/binder/BatchConsumerConfiguration.java
@@ -11,22 +11,26 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.MessageBuilder;
 
+import java.util.List;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 @Configuration
-@Profile("!manual")
-public class EventProducerConfiguration {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(EventHubBinderApplication.class);
-
+@Profile("batch")
+public class BatchConsumerConfiguration {
+    private static final Logger LOGGER = LoggerFactory.getLogger(BatchConsumerConfiguration.class);
     private int i = 0;
+
+    @Bean
+    public Consumer<List<String>> consume() {
+        return list -> list.forEach(event -> LOGGER.info("New event received: '{}'",event));
+    }
 
     @Bean
     public Supplier<Message<String>> supply() {
         return () -> {
             LOGGER.info("Sending message, sequence " + i);
-            return MessageBuilder.withPayload("Hello world, " + i++).build();
+            return MessageBuilder.withPayload("\"Hello world, " + i++ + "\"").build();
         };
     }
-
 }
