+++ eventhubs/azure-spring-cloud-stream-binder-eventhubs/eventhubs-binder/src/main/java/com/azure/spring/sample/eventhubs/binder/EventHubBinderApplication.java
@@ -3,19 +3,12 @@
 
 package com.azure.spring.sample.eventhubs.binder;
 
-import com.azure.spring.integration.core.EventHubHeaders;
-import com.azure.spring.integration.core.api.reactor.Checkpointer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.messaging.Message;
-
-import java.util.function.Consumer;
-
-import static com.azure.spring.integration.core.AzureHeaders.CHECKPOINTER;
 
 /**
  * @author Warren Zhu
@@ -27,25 +20,6 @@ public class EventHubBinderApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(EventHubBinderApplication.class, args);
-    }
-
-    @Bean
-    public Consumer<Message<String>> consume() {
-        return message -> {
-            Checkpointer checkpointer = (Checkpointer) message.getHeaders().get(CHECKPOINTER);
-            LOGGER.info("New message received: '{}', partition key: {}, sequence number: {}, offset: {}, enqueued time: {}",
-                message.getPayload(),
-                message.getHeaders().get(EventHubHeaders.PARTITION_KEY),
-                message.getHeaders().get(EventHubHeaders.SEQUENCE_NUMBER),
-                message.getHeaders().get(EventHubHeaders.OFFSET),
-                message.getHeaders().get(EventHubHeaders.ENQUEUED_TIME)
-                );
-
-            checkpointer.success()
-                        .doOnSuccess(success -> LOGGER.info("Message '{}' successfully checkpointed", message.getPayload()))
-                        .doOnError(error -> LOGGER.error("Exception found", error))
-                        .subscribe();
-        };
     }
 
     // Replace destination with spring.cloud.stream.bindings.consume-in-0.destination
