+++ eventhubs/azure-spring-cloud-stream-binder-eventhubs/eventhubs-binder/src/main/resources/application-batch.yaml
@@ -1,0 +1,33 @@
+spring:
+  cloud:
+    azure:
+      eventhub:
+        connection-string: [eventhub-namespace-connection-string]
+        checkpoint-storage-account: [checkpoint-storage-account]
+        checkpoint-access-key: [checkpoint-access-key]
+        checkpoint-container: [checkpoint-container]
+
+    stream:
+      function:
+        definition: consume;supply
+      bindings:
+        consume-in-0:
+          destination: [eventhub-name]
+          group: [consumer-group]
+          consumer:
+            batch-mode: true
+        supply-out-0:
+          destination: [eventhub-name]
+
+      eventhub:
+        bindings:
+          consume-in-0:
+            consumer:
+              checkpoint-mode: BATCH
+              max-batch-size: 5
+      default:
+        producer:
+          errorChannelEnabled: true
+      poller:
+        initial-delay: 0
+        fixed-delay: 1000
